requirements.txt
Django==4.2.4
django-browser-reload==1.11.0
django-import-export==3.2.0
django-tailwind==3.6.0
djangorestframework==3.14.0
django-simple-menu
